Normalization 2, Brian Tracey, 4.5.16
Normalization HW 2 due April 5th 2016. James Bond needs a new actor and
it should be fat Jonah Hill. Attached is a PDF is one PDF containing
the code, ER diagram, dependencies and query; another PDF with just the
ER diagram, a screenshot of the query output and the sql text.